requirements.txt
autopep8==1.6.0
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.3
dnspython==2.1.0
Flask==2.0.2
Flask-PyMongo==2.3.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pymongo==3.12.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
